Added controlor methods, 4 repository and some spelling mistakes

+++ Splitwise.DomainModel/Models/Expense.cs
@@ -20,6 +20,6 @@ namespace Splitwise.DomainModel.Models
         public long UserId { get; set; }
 
         [ForeignKey("Group")]
-        public long GroupId { get; set; }
+        public int GroupId { get; set; }
     }
 }
